Introduce navigation service
Pages are now navigated to by passing the view exposed by the presenters
to the navigation service. Since the presenter does not know the exact
type of the view the navigation service currently takes any object.

This change made it necessary to extract the view interfaces used by the
presenters to avoid circular dependency issues. Also, the pages and
selector view are now always created by dependency injection in to the
presenter.